Added in Authentication with Google. No testing done yet.

+++ controllers/diaryController.js
@@ -4,7 +4,7 @@ import { fetchWeather } from "./weatherController.js";
 export const getAllEntries = async(req, res) => {
     try{
         const { search, tag, location } = req.query;
-        let filter = {}; // No authentication in Part 1, so no user filter is applied
+        let filter = {user: req.user._id}; // No authentication in Part 1, so no user filter is applied
         // Search filter (Matches title or content)
         if (search) {
             filter.$or = [
@@ -35,6 +35,10 @@ export const getEntryById = async(req, res) => {
         if (!entry) {
             return res.status(404).json({ message: "Diary entry not found" });
         }
+
+        if (entry.user.toString() !== req.user._id.toString()) {
+            return res.status(403).json({ message: "Forbidden" });
+        }
         
         res.status(200).json(entry);
     }catch(error){
@@ -48,7 +52,7 @@ export const createEntry = async (req, res) => {
         // Fetch weather data if location is provided
         const weatherData = location ? await fetchWeather(location) : null;
         const newEntry = new DiaryEntry({
-            user: req.user?.id || "65f1234567890abcdef12345", // authentication is added in Part 2
+            user: req.user._id,
             title,
             content,
             reflection,
@@ -72,8 +76,8 @@ export const updateEntry = async(req, res) => {
             return res.status(404).json({ message: "Diary entry not found" });
         }
 
-        if (entry.user.toString() !== (req.user?.id || "65f1234567890abcdef12345")) {
-            return res.status(403).json({ message: "Not authorized to update this entry" });
+        if (entry.user.toString() !== req.user._id.toString()) {
+            return res.status(403).json({ message: "Forbidden" });
         }
 
         const updatedEntry = await DiaryEntry.findByIdAndUpdate(req.params.id, updates, {
@@ -90,13 +94,18 @@ export const updateEntry = async(req, res) => {
 
 export const deleteEntry = async(req, res) => {
     try{
-        const entry = await DiaryEntry.findByIdAndDelete(req.params.id);
-
+        const entry = await DiaryEntry.findById(req.params.id);
         if (!entry) {
             return res.status(404).json({ message: "Diary entry not found" });
         }
 
-        res.status(200).json(entry);
+        if (entry.user.toString() !== req.user._id.toString()) {
+            return res.status(403).json({ message: "Forbidden" });
+        }
+
+        const deletedEntry = await DiaryEntry.findByIdAndDelete(req.params.id);
+
+        res.status(200).json(deletedEntry);
     }catch(error){
         res.status(400).json({ message: error.message });
     }
